Implement System Environment Extension
This commit implements the System Environment Extensions. With some extension functions and a listener, this commit enables System Environment Override by users, something only possible via reflection.

An special attention is needed to the system environments, as they're susceptible to race conditions. This was specified in the documentations.

This commit also fixes a small issue with SystemProperties extension, in which they wouldn't accept a suspended function. A test for that behavior was added.

This commit also renames the tests package, to stop stacktrace filtering, so that the error is clear whenever one of these tests fail.

Implements #623